test: move SkipTestIfNoEnvVars to the correct place: package help
